test(docker_registry): refactor
